refactor: 테스트 폴더 구조 컨벤션 통일
